refactor the whole application
