+++ src/components/charts/Activity/Activity.css
