+++ src/components/charts/AverageSessions/AverageSessions.css
