+++ src/pages/Profile.jsx
@@ -1,10 +1,13 @@
 import Header from '@/components/Header/Header'
+import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
-import Activity from '../components/Activity/Activity'
-import AverageSessions from '../components/AverageSessions/AverageSessions'
-import KeyDatas from '../components/Keydatas/KeyDatas'
-import Performances from '../components/Performances/Performances'
-import Score from '../components/Score/Score'
+import { getUserData } from '../api/getUserData'
+import Loader from '../components/Loader/Loader'
+import Performances from '../components/charts/Performances/Performances'
+import Score from '../components/charts/Score/Score'
+import Activity from '../components/charts/Activity/Activity'
+import AverageSessions from '../components/charts/AverageSessions/AverageSessions'
+import KeyDatas from '../components/charts/Keydatas/KeyDatas'
 import './Profile.css'
 
 /**
@@ -14,6 +17,26 @@ import './Profile.css'
  */
 export default function Profile() {
     const { id } = useParams()
+    const [isLoading, setIsLoading] = useState(true)
+
+    useEffect(() => {
+        const fetchData = async () => {
+            try {
+                await getUserData(Number(id))
+            } catch (error) {
+                console.error(error)
+            } finally {
+                setIsLoading(false)
+            }
+        }
+
+        fetchData()
+    }, [id])
+
+    if (isLoading) {
+        return <Loader /> // Remplacez par votre composant de chargement
+    }
+
     return (
         <main>
             <div className="profil">
